Applying Stash
