+++ GrowApp/Features/AddEditPlantViewController/AddEditPlantViewControllerDelegate.swift
@@ -1,0 +1,18 @@
+//
+//  PlantConfigurationDelegate.swift
+//  GrowApp
+//
+//  Created by Ryan Thally on 4/2/21.
+//
+
+import Foundation
+
+protocol AddEditPlantViewControllerDelegate: AnyObject {
+    func plantEditor(_ editor: AddEditPlantViewController, didUpdatePlant plant: GHPlant)
+    func plantEditorDidCancel(_ editor: AddEditPlantViewController)
+}
+
+extension AddEditPlantViewControllerDelegate {
+    func plantEditor(_ editor: AddEditPlantViewController, didUpdatePlant plant: GHPlant) { }
+    func plantEditorDidCancel(_ editor: AddEditPlantViewController) { }
+}
