--- GrowApp/Features/Plant Editor/TaskEditor/TaskEditorDelegate.swift
@@ -1,18 +1,0 @@
-//
-//  TaskEditorDelegate.swift
-//  GrowApp
-//
-//  Created by Ryan Thally on 4/9/21.
-//
-
-import Foundation
-
-protocol TaskEditorDelegate {
-    func taskEditor(_ editor: TaskEditorController, didUpdateTask task: CareInfo)
-    func taskEditorDidCancel(_ editor: TaskEditorController)
-}
-
-extension TaskEditorDelegate {
-    func taskEditor(_ editor: TaskEditorController, didUpdateTask task: CareInfo) { }
-    func taskEditorDidCancel(_ editor: TaskEditorController) { }
-}
